+++ R/utils.R
@@ -1054,7 +1054,7 @@ makeopts <- function(method = 'MHRM', draws = 2000L, calcLL = TRUE, quadpts = NU
     tnames <- names(technical)
     gnames <- c('MAXQUAD', 'NCYCLES', 'BURNIN', 'SEMCYCLES', 'set.seed', 'SEtol', 'symmetric_SEM',
                 'gain', 'warn', 'message', 'customK', 'customPriorFun', 'customTheta', 'MHcand',
-                'parallel', 'NULL.MODEL', 'theta_lim', 'RANDSTART', 'MHDRAWS')
+                'parallel', 'NULL.MODEL', 'theta_lim', 'RANDSTART', 'MHDRAWS', 'removeEmptyRows')
     if(!all(tnames %in% gnames))
         stop('The following inputs to technical are invalid: ',
              paste0(tnames[!(tnames %in% gnames)], ' '), call.=FALSE)
@@ -1087,6 +1087,8 @@ makeopts <- function(method = 'MHRM', draws = 2000L, calcLL = TRUE, quadpts = NU
         if(is.null(technical$NCYCLES)) technical$NCYCLES <- 1000L
     }
     if(is.null(technical$symmetric_SEM)) technical$symmetric_SEM <- TRUE
+    opts$removeEmptyRows <- if(is.null(technical$removeEmptyRows)) FALSE
+        else technical$removeEmptyRows
     opts$warn <- if(is.null(technical$warn)) TRUE else technical$warn
     opts$message <- if(is.null(technical$message)) TRUE else technical$message
     opts$technical <- technical